+++ spec/lib/heidrun/mapping_tools_spec.rb
@@ -1,0 +1,4 @@
+require 'spec_helper'
+
+describe Heidrun::MappingTools do
+end